fix: AppflyerAppDelegateService
