+++ p2p_wallet/Info.plist
@@ -97,11 +97,6 @@
 	<string>$(GOOGLE_SIGN_IN_ID_RELEASE)</string>
 	<key>ITSAppUsesNonExemptEncryption</key>
 	<false/>
-	<key>IntercomUniversalLinkDomains</key>
-	<array>
-		<string>t.key.app</string>
-		<string>key.app</string>
-	</array>
 	<key>LSRequiresIPhoneOS</key>
 	<true/>
 	<key>MOONPAY_PRODUCTION_API_KEY</key>
