+++ p2p_wallet/Resources/AppDelegate/AppDelegateProxyService.swift
@@ -13,6 +13,7 @@ final class AppDelegateProxyService: NSObject, UIApplicationDelegate {
 
     override init() {
         var services: [AppDelegateService] = [
+            AppflyerAppDelegateService(),
             DeeplinkAppDelegateService()
         ]
         #if !RELEASE
