+++ p2p_wallet/Resources/AppDelegate/DeeplinkAppDelegateService.swift
@@ -9,10 +9,7 @@ import Foundation
 import Resolver
 
 final class DeeplinkAppDelegateService: NSObject, AppDelegateService {
-    @Injected var userWalletManager: UserWalletManager
-    @Injected var appEventHandler: AppEventHandlerType
-    @Injected var pincodeStorageService: PincodeStorageType
-    @Injected var authService: AuthenticationHandlerType
+    // MARK: - URIScheme
 
     func application(_ app: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey: Any] = [:]) -> Bool {
         guard
@@ -29,6 +26,10 @@ final class DeeplinkAppDelegateService: NSObject, AppDelegateService {
             let seedPhrase: String = params.first(where: { $0.name == "value" })?.value,
             let pincode: String = params.first(where: { $0.name == "pincode" })?.value
         {
+            let userWalletManager: UserWalletManager = Resolver.resolve()
+            let appEventHandler: AppEventHandlerType = Resolver.resolve()
+            let pincodeStorageService: PincodeStorageType = Resolver.resolve()
+            let authService: AuthenticationHandlerType = Resolver.resolve()
             Task {
                 appEventHandler.delegate?.disablePincodeOnFirstAppear()
                 pincodeStorageService.save(pincode)
@@ -42,6 +43,34 @@ final class DeeplinkAppDelegateService: NSObject, AppDelegateService {
             return true
         }
 
+        return false
+    }
+    
+    // MARK: - Universal links
+    
+    func application(_ application: UIApplication, continue userActivity: NSUserActivity, restorationHandler: @escaping ([UIUserActivityRestoring]?) -> Void) -> Bool {
+        if
+            let webpageURL = userActivity.webpageURL,
+            let urlComponents = URLComponents(url: webpageURL, resolvingAgainstBaseURL: true)
+        {
+            // Intercom survey
+            if urlComponents.path == "/intercom",
+               let queryItem = urlComponents.queryItems?.first(where: { $0.name == "intercom_survey_id" }),
+               let value = queryItem.value
+            {
+                DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
+                    GlobalAppState.shared.surveyID = value
+                }
+                return true
+            }
+            
+            // send via link
+            else if webpageURL.host == "t.key.app" {
+                GlobalAppState.shared.sendViaLinkUrl = webpageURL
+                return true
+            }
+        }
+        
         return false
     }
 }
